requirements.txt
# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # Commented out for SQLite demo

# Background Tasks (commented out for demo)
# celery==5.3.6
# redis==5.0.1

# Code Analysis Tools
pylint==3.0.3
bandit[toml]==1.7.6
radon==6.0.1
coverage==7.4.0
black==24.1.1
mypy==1.8.0
flake8==7.0.0
flake8-bugbear==24.1.17  # Additional flake8 plugin for bug detection
flake8-security==1.7.1   # Security-focused linting
safety==3.0.1
vulture==2.11

# Git Operations
GitPython==3.1.41

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-xdist==3.5.0      # Parallel test execution

# Security & Authentication
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0     # Explicit version for security

# Utilities
python-dotenv==1.0.1
structlog==23.2.0
